make MacroTools.@q as development-only dependency
