+++ TyPython/src/CPython.Dev.DevOnly.compiled.jl
@@ -2,8 +2,8 @@
 import Serialization
 import SHA
 let
-    expected_sign = b"\x21\x3c\x66\x60\xf2\x5b\xa3\x2b\x0c\xa0\x0a\xac\xc5\x53\xe4\x85\xd7\x18\x2b\x65\x85\x4b\x7c\x65\x8e\xb3\x21\xa6"
-    data = b"\x37\x4a\x4c\x0f\x04\x00\x00\x00\x16\x02\x01\x08\x74\x6f\x70\x6c\x65\x76\x65\x6c\x17\xe0\x01\x3b\x45\x3a\x5c\x50\x72\x6f\x6a\x65\x63\x74\x73\x5c\x47\x69\x74\x34\x5c\x6a\x6e\x75\x6d\x70\x79\x5c\x54\x79\x50\x79\x74\x68\x6f\x6e\x5c\x73\x72\x63\x5c\x43\x50\x79\x74\x68\x6f\x6e\x2e\x44\x65\x76\x2e\x44\x65\x76\x4f\x6e\x6c\x79\x2e\x6a\x6c\x16\x02\x01\x05\x6d\x61\x63\x72\x6f\x16\x03\x56\x01\x0f\x65\x78\x70\x6f\x72\x74\x5f\x70\x79\x6d\x6f\x64\x75\x6c\x65\x16\x02\x57\x01\x04\x6e\x61\x6d\x65\x01\x06\x53\x79\x6d\x62\x6f\x6c\x01\x02\x65\x78\x16\x13\x01\x05\x62\x6c\x6f\x63\x6b\x17\xe0\x2c\x02\x00\x17\xe1\x2c\x02\x00\x16\x09\x01\x05\x62\x6c\x6f\x63\x6b\x16\x01\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x0b\x23\x23\x63\x61\x63\x68\x65\x23\x37\x36\x30\x01\x07\x6e\x6f\x74\x68\x69\x6e\x67\x16\x02\x59\x01\x05\x23\x23\x37\x35\x39\x01\x02\x65\x78\x17\xe2\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x35\x39\x00\x16\x16\x02\x01\x02\x26\x26\x16\x03\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x01\x02\x69\x66\x16\x03\x56\x5b\x2c\x0b\x00\x01\x07\x6e\x6f\x74\x68\x69\x6e\x67\x16\x02\x59\x2c\x0b\x00\x16\x02\x56\x01\x04\x53\x6f\x6d\x65\x16\x02\xa4\x16\x02\x01\x01\x2e\x01\x05\x23\x23\x37\x35\x39\x1a\x01\x04\x68\x65\x61\x64\x16\x02\x01\x01\x2e\x01\x05\x23\x23\x37\x35\x39\x1a\x01\x04\x61\x72\x67\x73\x16\x02\x59\x01\x05\x23\x23\x37\x36\x31\x16\x02\x01\x01\x2e\x2c\x0b\x00\x1a\x01\x05\x76\x61\x6c\x75\x65\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x36\x31\x12\x00\x1c\x01\x04\x76\x61\x72\x32\x47\x12\x01\x02\x00\x10\x01\x0d\x41\x62\x73\x74\x72\x61\x63\x74\x41\x72\x72\x61\x79\x1f\x4e\x9b\x44\x00\x00\x00\x00\x10\x53\x1f\x4e\x9b\x44\x02\x00\x00\x00\x00\x01\x2c\x1c\x00\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5a\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x36\x31\xe0\x1a\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x01\x02\x26\x26\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x05\x23\x23\x37\x36\x32\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x36\x31\xe1\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x36\x32\x12\x01\x02\x00\x2c\x1d\x00\x16\x02\x01\x02\x26\x26\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x56\x01\x05\x6e\x64\x69\x6d\x73\x01\x05\x23\x23\x37\x36\x32\xe0\x16\x03\x56\x01\x02\x3e\x3d\x16\x02\x56\x01\x06\x6c\x65\x6e\x67\x74\x68\x01\x05\x23\x23\x37\x36\x32\xdf\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x05\x23\x23\x37\x36\x33\x16\x03\x56\x12\x01\x05\x00\x10\x01\x08\x53\x75\x62\x41\x72\x72\x61\x79\x1f\x4e\x9e\x44\x00\x00\x00\x00\x01\x05\x23\x23\x37\x36\x32\x16\x01\xa4\x16\x03\x56\x01\x01\x3a\xe0\x16\x02\x56\x01\x06\x6c\x65\x6e\x67\x74\x68\x01\x05\x23\x23\x37\x36\x32\x4d\x16\x03\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x05\x73\x75\x69\x74\x65\x01\x05\x23\x23\x37\x36\x33\x16\x02\x59\x01\x0c\x23\x23\x72\x65\x74\x75\x72\x6e\x23\x37\x35\x37\x16\x00\x01\x05\x62\x6c\x6f\x63\x6b\x16\x01\x01\x0c\x73\x79\x6d\x62\x6f\x6c\x69\x63\x67\x6f\x74\x6f\x01\x11\x23\x23\x23\x23\x66\x69\x6e\x61\x6c\x23\x37\x35\x38\x23\x37\x36\x34\x17\xe3\x2c\x02\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x39\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xe4\x2c\x02\x00\x16\x02\x56\x01\x05\x65\x72\x72\x6f\x72\x39\x21\x2a\x40\x65\x78\x70\x6f\x72\x74\x5f\x70\x79\x6d\x6f\x64\x75\x6c\x65\x20\x65\x78\x70\x65\x63\x74\x73\x20\x61\x20\x62\x65\x67\x69\x6e\x2d\x65\x6e\x64\x20\x62\x6c\x6f\x63\x6b\x16\x01\x2c\x3c\x00\x2c\x3d\x00\x16\x02\x56\x34\x10\x01\x06\x23\x65\x72\x72\x6f\x72\x1f\x4e\x9e\x44\x39\x21\x5f\x6d\x61\x74\x63\x68\x69\x6e\x67\x20\x6e\x6f\x6e\x2d\x65\x78\x68\x61\x75\x73\x74\x69\x76\x65\x2c\x20\x61\x74\x20\x23\x3d\x20\x45\x3a\x5c\x50\x72\x6f\x6a\x65\x63\x74\x73\x5c\x47\x69\x74\x34\x5c\x6a\x6e\x75\x6d\x70\x79\x5c\x54\x79\x50\x79\x74\x68\x6f\x6e\x5c\x73\x72\x63\x5c\x43\x50\x79\x74\x68\x6f\x6e\x2e\x44\x65\x76\x2e\x44\x65\x76\x4f\x6e\x6c\x79\x2e\x6a\x6c\x3a\x32\x20\x3d\x23\x16\x01\x01\x0d\x73\x79\x6d\x62\x6f\x6c\x69\x63\x6c\x61\x62\x65\x6c\x2c\x3d\x00\x2c\x39\x00\x17\xe6\x2c\x02\x00\x16\x02\x59\x01\x04\x62\x6f\x64\x79\x16\x02\x56\x01\x04\x45\x78\x70\x72\x1a\x01\x05\x62\x6c\x6f\x63\x6b\x17\xe7\x2c\x02\x00\x16\x02\x59\x01\x03\x6f\x75\x74\x16\x04\x56\x01\x04\x45\x78\x70\x72\x1a\x01\x03\x6c\x65\x74\x16\x02\x56\x01\x04\x45\x78\x70\x72\x1a\x01\x05\x62\x6c\x6f\x63\x6b\x01\x04\x62\x6f\x64\x79\x17\xe8\x2c\x02\x00\x16\x02\x59\x01\x0a\x73\x79\x6d\x5f\x6d\x6f\x64\x75\x6c\x65\x16\x02\x56\x01\x06\x67\x65\x6e\x73\x79\x6d\x16\x02\x01\x06\x73\x74\x72\x69\x6e\x67\x21\x04\x6d\x6f\x64\x5f\x01\x04\x6e\x61\x6d\x65\x17\xe9\x2c\x02\x00\x16\x02\x59\x01\x0b\x6d\x6f\x64\x75\x6c\x65\x5f\x6e\x61\x6d\x65\x16\x02\x56\x01\x06\x73\x74\x72\x69\x6e\x67\x01\x04\x6e\x61\x6d\x65\x17\xea\x2c\x02\x00\x16\x03\x56\x01\x05\x70\x75\x73\x68\x21\x16\x02\x01\x01\x2e\x01\x04\x62\x6f\x64\x79\x1a\x01\x04\x61\x72\x67\x73\x16\x03\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x05\x6c\x6f\x63\x61\x6c\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x59\x2c\x4f\x00\x16\x03\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x01\x07\x43\x50\x79\x74\x68\x6f\x6e\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x1a\x01\x0b\x47\x5f\x50\x79\x42\x75\x69\x6c\x74\x69\x6e\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x1a\x01\x0a\x5f\x5f\x69\x6d\x70\x6f\x72\x74\x5f\x5f\x16\x05\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x01\x07\x70\x79\x5f\x63\x61\x73\x74\x01\x02\x50\x79\x21\x05\x74\x79\x70\x65\x73\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x1a\x01\x0f\x53\x69\x6d\x70\x6c\x65\x4e\x61\x6d\x65\x73\x70\x61\x63\x65\x17\xeb\x2c\x02\x00\x16\x02\x01\x03\x66\x6f\x72\x16\x02\x59\x01\x03\x61\x72\x67\x01\x05\x73\x75\x69\x74\x65\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xec\x2c\x02\x00\x16\x09\x01\x05\x62\x6c\x6f\x63\x6b\x16\x01\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x0b\x23\x23\x63\x61\x63\x68\x65\x23\x37\x36\x38\x01\x07\x6e\x6f\x74\x68\x69\x6e\x67\x16\x02\x59\x01\x05\x23\x23\x37\x36\x37\x01\x03\x61\x72\x67\x16\x02\x01\x02\x69\x66\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x36\x37\x00\x16\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xed\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x01\x02\x69\x66\x16\x03\x56\x5b\x2c\x6b\x00\x01\x07\x6e\x6f\x74\x68\x69\x6e\x67\x16\x02\x59\x2c\x6b\x00\x16\x02\x56\x01\x04\x53\x6f\x6d\x65\x16\x02\xa4\x16\x02\x01\x01\x2e\x01\x05\x23\x23\x37\x36\x37\x1a\x01\x04\x68\x65\x61\x64\x16\x02\x01\x01\x2e\x01\x05\x23\x23\x37\x36\x37\x1a\x01\x04\x61\x72\x67\x73\x16\x02\x59\x01\x05\x23\x23\x37\x36\x39\x16\x02\x01\x01\x2e\x2c\x6b\x00\x1a\x01\x05\x76\x61\x6c\x75\x65\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x36\x39\x12\x00\x1c\x01\x04\x76\x61\x72\x32\x47\x12\x01\x02\x00\x2c\x1d\x00\x10\x53\x1f\x4e\x9b\x44\x02\x00\x00\x00\x00\x01\x2c\x7d\x00\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5a\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x36\x39\xe0\x1a\x59\x16\x02\x01\x02\x26\x26\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x05\x23\x23\x37\x37\x30\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x36\x39\xe1\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x37\x30\x12\x01\x02\x00\x2c\x1d\x00\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x56\x01\x06\x6c\x65\x6e\x67\x74\x68\x01\x05\x23\x23\x37\x37\x30\xe1\x16\x03\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x05\x23\x23\x37\x37\x31\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x37\x30\xe0\x16\x02\x59\x01\x05\x23\x23\x37\x37\x32\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x37\x30\xe1\x4d\x16\x04\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x05\x76\x61\x6c\x75\x65\x01\x05\x23\x23\x37\x37\x32\x16\x02\x59\x01\x04\x6e\x61\x6d\x65\x01\x05\x23\x23\x37\x37\x31\x16\x02\x59\x01\x0c\x23\x23\x72\x65\x74\x75\x72\x6e\x23\x37\x36\x35\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xee\x2c\x02\x00\x16\x03\x56\x01\x05\x70\x75\x73\x68\x21\x16\x02\x01\x01\x2e\x01\x04\x62\x6f\x64\x79\x1a\x01\x04\x61\x72\x67\x73\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x59\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x2c\x4f\x00\x16\x02\x56\x24\x1f\x4e\x9b\x44\x01\x09\x51\x75\x6f\x74\x65\x4e\x6f\x64\x65\x01\x04\x6e\x61\x6d\x65\x16\x05\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x01\x07\x70\x79\x5f\x63\x61\x73\x74\x01\x02\x50\x79\x01\x05\x76\x61\x6c\x75\x65\x16\x01\x2c\x3c\x00\x01\x11\x23\x23\x23\x23\x66\x69\x6e\x61\x6c\x23\x37\x36\x36\x23\x37\x37\x33\x16\x02\x01\x02\x69\x66\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x36\x37\x00\x17\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xef\x2c\x02\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x93\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xf0\x2c\x02\x00\x16\x03\x56\x01\x05\x70\x75\x73\x68\x21\x16\x02\x01\x01\x2e\x01\x04\x62\x6f\x64\x79\x1a\x01\x04\x61\x72\x67\x73\x01\x03\x61\x72\x67\x16\x01\x2c\x3c\x00\x2c\x9d\x00\x17\xf1\x2c\x02\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x93\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xf2\x2c\x02\x00\x16\x02\x56\x01\x05\x65\x72\x72\x6f\x72\x39\x21\x3d\x40\x65\x78\x70\x6f\x72\x74\x5f\x70\x79\x6d\x6f\x64\x75\x6c\x65\x20\x65\x78\x70\x65\x63\x74\x73\x20\x61\x20\x62\x6c\x6f\x63\x6b\x20\x6f\x66\x20\x60\x6e\x61\x6d\x65\x20\x3d\x20\x76\x61\x6c\x75\x65\x60\x20\x73\x74\x61\x74\x65\x6d\x65\x6e\x74\x73\x16\x01\x2c\x3c\x00\x2c\x9d\x00\x16\x02\x56\x34\x2c\x45\x00\x39\x21\x60\x6d\x61\x74\x63\x68\x69\x6e\x67\x20\x6e\x6f\x6e\x2d\x65\x78\x68\x61\x75\x73\x74\x69\x76\x65\x2c\x20\x61\x74\x20\x23\x3d\x20\x45\x3a\x5c\x50\x72\x6f\x6a\x65\x63\x74\x73\x5c\x47\x69\x74\x34\x5c\x6a\x6e\x75\x6d\x70\x79\x5c\x54\x79\x50\x79\x74\x68\x6f\x6e\x5c\x73\x72\x63\x5c\x43\x50\x79\x74\x68\x6f\x6e\x2e\x44\x65\x76\x2e\x44\x65\x76\x4f\x6e\x6c\x79\x2e\x6a\x6c\x3a\x31\x33\x20\x3d\x23\x16\x01\x2c\x48\x00\x2c\x9d\x00\x2c\x93\x00\x17\xf5\x2c\x02\x00\x16\x03\x56\x01\x05\x70\x75\x73\x68\x21\x16\x02\x01\x01\x2e\x01\x04\x62\x6f\x64\x79\x1a\x01\x04\x61\x72\x67\x73\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x59\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x03\x72\x65\x66\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x01\x07\x43\x50\x79\x74\x68\x6f\x6e\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x1a\x2c\x5f\x00\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x1a\x2c\x61\x00\x16\x05\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x01\x07\x70\x79\x5f\x63\x61\x73\x74\x01\x02\x50\x79\x21\x03\x73\x79\x73\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x1a\x01\x07\x6d\x6f\x64\x75\x6c\x65\x73\x16\x05\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x01\x07\x70\x79\x5f\x63\x61\x73\x74\x01\x02\x50\x79\x2c\x53\x00\x2c\x4f\x00\x17\xf6\x2c\x02\x00\x16\x02\x56\x01\x03\x65\x73\x63\x01\x03\x6f\x75\x74"
+    expected_sign = b"\xe2\xd6\xd2\xd0\xca\x7c\xf8\x2b\x76\xc3\x86\x8e\x4e\x95\x26\xe1\x17\x49\x88\x43\x1a\x94\x94\x59\x6d\x3d\xbe\x6b"
+    data = b"\x37\x4a\x4c\x0d\x04\x00\x00\x00\x16\x02\x01\x08\x74\x6f\x70\x6c\x65\x76\x65\x6c\x17\xe0\x01\x3b\x45\x3a\x5c\x50\x72\x6f\x6a\x65\x63\x74\x73\x5c\x47\x69\x74\x34\x5c\x6a\x6e\x75\x6d\x70\x79\x5c\x54\x79\x50\x79\x74\x68\x6f\x6e\x5c\x73\x72\x63\x5c\x43\x50\x79\x74\x68\x6f\x6e\x2e\x44\x65\x76\x2e\x44\x65\x76\x4f\x6e\x6c\x79\x2e\x6a\x6c\x16\x02\x01\x05\x6d\x61\x63\x72\x6f\x16\x03\x56\x01\x0f\x65\x78\x70\x6f\x72\x74\x5f\x70\x79\x6d\x6f\x64\x75\x6c\x65\x16\x02\x57\x01\x04\x6e\x61\x6d\x65\x01\x06\x53\x79\x6d\x62\x6f\x6c\x01\x02\x65\x78\x16\x13\x01\x05\x62\x6c\x6f\x63\x6b\x17\xe0\x2c\x02\x00\x17\xe1\x2c\x02\x00\x16\x09\x01\x05\x62\x6c\x6f\x63\x6b\x16\x01\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x0b\x23\x23\x63\x61\x63\x68\x65\x23\x37\x34\x34\x01\x07\x6e\x6f\x74\x68\x69\x6e\x67\x16\x02\x59\x01\x05\x23\x23\x37\x34\x33\x01\x02\x65\x78\x17\xe2\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x34\x33\x00\x16\x16\x02\x01\x02\x26\x26\x16\x03\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x01\x02\x69\x66\x16\x03\x56\x5b\x2c\x0b\x00\x01\x07\x6e\x6f\x74\x68\x69\x6e\x67\x16\x02\x59\x2c\x0b\x00\x16\x02\x56\x01\x04\x53\x6f\x6d\x65\x16\x02\xa4\x16\x02\x01\x01\x2e\x01\x05\x23\x23\x37\x34\x33\x1a\x01\x04\x68\x65\x61\x64\x16\x02\x01\x01\x2e\x01\x05\x23\x23\x37\x34\x33\x1a\x01\x04\x61\x72\x67\x73\x16\x02\x59\x01\x05\x23\x23\x37\x34\x35\x16\x02\x01\x01\x2e\x2c\x0b\x00\x1a\x01\x05\x76\x61\x6c\x75\x65\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x34\x35\x12\x00\x1c\x01\x04\x76\x61\x72\x32\x47\x12\x01\x02\x00\x10\x01\x0d\x41\x62\x73\x74\x72\x61\x63\x74\x41\x72\x72\x61\x79\x1f\x4e\x9b\x44\x00\x00\x00\x00\x10\x53\x1f\x4e\x9b\x44\x02\x00\x00\x00\x00\x01\x2c\x1c\x00\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5a\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x34\x35\xe0\x1a\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x01\x02\x26\x26\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x05\x23\x23\x37\x34\x36\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x34\x35\xe1\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x34\x36\x12\x01\x02\x00\x2c\x1d\x00\x16\x02\x01\x02\x26\x26\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x56\x01\x05\x6e\x64\x69\x6d\x73\x01\x05\x23\x23\x37\x34\x36\xe0\x16\x03\x56\x01\x02\x3e\x3d\x16\x02\x56\x01\x06\x6c\x65\x6e\x67\x74\x68\x01\x05\x23\x23\x37\x34\x36\xdf\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x05\x23\x23\x37\x34\x37\x16\x03\x56\x12\x01\x05\x00\x10\x01\x08\x53\x75\x62\x41\x72\x72\x61\x79\x1f\x4e\x9e\x44\x00\x00\x00\x00\x01\x05\x23\x23\x37\x34\x36\x16\x01\xa4\x16\x03\x56\x01\x01\x3a\xe0\x16\x02\x56\x01\x06\x6c\x65\x6e\x67\x74\x68\x01\x05\x23\x23\x37\x34\x36\x4d\x16\x03\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x05\x73\x75\x69\x74\x65\x01\x05\x23\x23\x37\x34\x37\x16\x02\x59\x01\x0c\x23\x23\x72\x65\x74\x75\x72\x6e\x23\x37\x34\x31\x16\x00\x01\x05\x62\x6c\x6f\x63\x6b\x16\x01\x01\x0c\x73\x79\x6d\x62\x6f\x6c\x69\x63\x67\x6f\x74\x6f\x01\x11\x23\x23\x23\x23\x66\x69\x6e\x61\x6c\x23\x37\x34\x32\x23\x37\x34\x38\x17\xe3\x2c\x02\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x39\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xe4\x2c\x02\x00\x16\x02\x56\x01\x05\x65\x72\x72\x6f\x72\x39\x21\x2a\x40\x65\x78\x70\x6f\x72\x74\x5f\x70\x79\x6d\x6f\x64\x75\x6c\x65\x20\x65\x78\x70\x65\x63\x74\x73\x20\x61\x20\x62\x65\x67\x69\x6e\x2d\x65\x6e\x64\x20\x62\x6c\x6f\x63\x6b\x16\x01\x2c\x3c\x00\x2c\x3d\x00\x16\x02\x56\x34\x10\x01\x06\x23\x65\x72\x72\x6f\x72\x1f\x4e\x9e\x44\x39\x21\x5f\x6d\x61\x74\x63\x68\x69\x6e\x67\x20\x6e\x6f\x6e\x2d\x65\x78\x68\x61\x75\x73\x74\x69\x76\x65\x2c\x20\x61\x74\x20\x23\x3d\x20\x45\x3a\x5c\x50\x72\x6f\x6a\x65\x63\x74\x73\x5c\x47\x69\x74\x34\x5c\x6a\x6e\x75\x6d\x70\x79\x5c\x54\x79\x50\x79\x74\x68\x6f\x6e\x5c\x73\x72\x63\x5c\x43\x50\x79\x74\x68\x6f\x6e\x2e\x44\x65\x76\x2e\x44\x65\x76\x4f\x6e\x6c\x79\x2e\x6a\x6c\x3a\x32\x20\x3d\x23\x16\x01\x01\x0d\x73\x79\x6d\x62\x6f\x6c\x69\x63\x6c\x61\x62\x65\x6c\x2c\x3d\x00\x2c\x39\x00\x17\xe6\x2c\x02\x00\x16\x02\x59\x01\x04\x62\x6f\x64\x79\x16\x02\x56\x01\x04\x45\x78\x70\x72\x1a\x01\x05\x62\x6c\x6f\x63\x6b\x17\xe7\x2c\x02\x00\x16\x02\x59\x01\x03\x6f\x75\x74\x16\x04\x56\x01\x04\x45\x78\x70\x72\x1a\x01\x03\x6c\x65\x74\x16\x02\x56\x01\x04\x45\x78\x70\x72\x1a\x01\x05\x62\x6c\x6f\x63\x6b\x01\x04\x62\x6f\x64\x79\x17\xe8\x2c\x02\x00\x16\x02\x59\x01\x0a\x73\x79\x6d\x5f\x6d\x6f\x64\x75\x6c\x65\x16\x02\x56\x01\x06\x67\x65\x6e\x73\x79\x6d\x16\x02\x01\x06\x73\x74\x72\x69\x6e\x67\x21\x04\x6d\x6f\x64\x5f\x01\x04\x6e\x61\x6d\x65\x17\xe9\x2c\x02\x00\x16\x02\x59\x01\x0b\x6d\x6f\x64\x75\x6c\x65\x5f\x6e\x61\x6d\x65\x16\x02\x56\x01\x06\x73\x74\x72\x69\x6e\x67\x01\x04\x6e\x61\x6d\x65\x17\xea\x2c\x02\x00\x16\x03\x56\x01\x05\x70\x75\x73\x68\x21\x16\x02\x01\x01\x2e\x01\x04\x62\x6f\x64\x79\x1a\x01\x04\x61\x72\x67\x73\x16\x03\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x05\x6c\x6f\x63\x61\x6c\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x59\x2c\x4f\x00\x16\x03\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x01\x07\x43\x50\x79\x74\x68\x6f\x6e\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x1a\x01\x0b\x47\x5f\x50\x79\x42\x75\x69\x6c\x74\x69\x6e\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x1a\x01\x0a\x5f\x5f\x69\x6d\x70\x6f\x72\x74\x5f\x5f\x16\x05\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x01\x07\x70\x79\x5f\x63\x61\x73\x74\x01\x02\x50\x79\x21\x05\x74\x79\x70\x65\x73\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x1a\x01\x0f\x53\x69\x6d\x70\x6c\x65\x4e\x61\x6d\x65\x73\x70\x61\x63\x65\x17\xeb\x2c\x02\x00\x16\x02\x01\x03\x66\x6f\x72\x16\x02\x59\x01\x03\x61\x72\x67\x01\x05\x73\x75\x69\x74\x65\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xec\x2c\x02\x00\x16\x09\x01\x05\x62\x6c\x6f\x63\x6b\x16\x01\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x0b\x23\x23\x63\x61\x63\x68\x65\x23\x37\x35\x32\x01\x07\x6e\x6f\x74\x68\x69\x6e\x67\x16\x02\x59\x01\x05\x23\x23\x37\x35\x31\x01\x03\x61\x72\x67\x16\x02\x01\x02\x69\x66\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x35\x31\x00\x16\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xed\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x01\x02\x69\x66\x16\x03\x56\x5b\x2c\x6b\x00\x01\x07\x6e\x6f\x74\x68\x69\x6e\x67\x16\x02\x59\x2c\x6b\x00\x16\x02\x56\x01\x04\x53\x6f\x6d\x65\x16\x02\xa4\x16\x02\x01\x01\x2e\x01\x05\x23\x23\x37\x35\x31\x1a\x01\x04\x68\x65\x61\x64\x16\x02\x01\x01\x2e\x01\x05\x23\x23\x37\x35\x31\x1a\x01\x04\x61\x72\x67\x73\x16\x02\x59\x01\x05\x23\x23\x37\x35\x33\x16\x02\x01\x01\x2e\x2c\x6b\x00\x1a\x01\x05\x76\x61\x6c\x75\x65\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x35\x33\x12\x00\x1c\x01\x04\x76\x61\x72\x32\x47\x12\x01\x02\x00\x2c\x1d\x00\x10\x53\x1f\x4e\x9b\x44\x02\x00\x00\x00\x00\x01\x2c\x7d\x00\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5a\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x35\x33\xe0\x1a\x59\x16\x02\x01\x02\x26\x26\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x05\x23\x23\x37\x35\x34\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x35\x33\xe1\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x35\x34\x12\x01\x02\x00\x2c\x1d\x00\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x56\x01\x06\x6c\x65\x6e\x67\x74\x68\x01\x05\x23\x23\x37\x35\x34\xe1\x16\x03\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x05\x23\x23\x37\x35\x35\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x35\x34\xe0\x16\x02\x59\x01\x05\x23\x23\x37\x35\x36\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x35\x34\xe1\x4d\x16\x04\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x05\x76\x61\x6c\x75\x65\x01\x05\x23\x23\x37\x35\x36\x16\x02\x59\x01\x04\x6e\x61\x6d\x65\x01\x05\x23\x23\x37\x35\x35\x16\x02\x59\x01\x0c\x23\x23\x72\x65\x74\x75\x72\x6e\x23\x37\x34\x39\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xee\x2c\x02\x00\x16\x03\x56\x01\x05\x70\x75\x73\x68\x21\x16\x02\x01\x01\x2e\x01\x04\x62\x6f\x64\x79\x1a\x01\x04\x61\x72\x67\x73\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x59\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x2c\x4f\x00\x16\x02\x56\x24\x1f\x4e\x9b\x44\x01\x09\x51\x75\x6f\x74\x65\x4e\x6f\x64\x65\x01\x04\x6e\x61\x6d\x65\x16\x05\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x01\x07\x70\x79\x5f\x63\x61\x73\x74\x01\x02\x50\x79\x01\x05\x76\x61\x6c\x75\x65\x16\x01\x2c\x3c\x00\x01\x11\x23\x23\x23\x23\x66\x69\x6e\x61\x6c\x23\x37\x35\x30\x23\x37\x35\x37\x16\x02\x01\x02\x69\x66\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x35\x31\x00\x17\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xef\x2c\x02\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x93\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xf0\x2c\x02\x00\x16\x03\x56\x01\x05\x70\x75\x73\x68\x21\x16\x02\x01\x01\x2e\x01\x04\x62\x6f\x64\x79\x1a\x01\x04\x61\x72\x67\x73\x01\x03\x61\x72\x67\x16\x01\x2c\x3c\x00\x2c\x9d\x00\x17\xf1\x2c\x02\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x93\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xf2\x2c\x02\x00\x16\x02\x56\x01\x05\x65\x72\x72\x6f\x72\x39\x21\x3d\x40\x65\x78\x70\x6f\x72\x74\x5f\x70\x79\x6d\x6f\x64\x75\x6c\x65\x20\x65\x78\x70\x65\x63\x74\x73\x20\x61\x20\x62\x6c\x6f\x63\x6b\x20\x6f\x66\x20\x60\x6e\x61\x6d\x65\x20\x3d\x20\x76\x61\x6c\x75\x65\x60\x20\x73\x74\x61\x74\x65\x6d\x65\x6e\x74\x73\x16\x01\x2c\x3c\x00\x2c\x9d\x00\x16\x02\x56\x34\x2c\x45\x00\x39\x21\x60\x6d\x61\x74\x63\x68\x69\x6e\x67\x20\x6e\x6f\x6e\x2d\x65\x78\x68\x61\x75\x73\x74\x69\x76\x65\x2c\x20\x61\x74\x20\x23\x3d\x20\x45\x3a\x5c\x50\x72\x6f\x6a\x65\x63\x74\x73\x5c\x47\x69\x74\x34\x5c\x6a\x6e\x75\x6d\x70\x79\x5c\x54\x79\x50\x79\x74\x68\x6f\x6e\x5c\x73\x72\x63\x5c\x43\x50\x79\x74\x68\x6f\x6e\x2e\x44\x65\x76\x2e\x44\x65\x76\x4f\x6e\x6c\x79\x2e\x6a\x6c\x3a\x31\x33\x20\x3d\x23\x16\x01\x2c\x48\x00\x2c\x9d\x00\x2c\x93\x00\x17\xf5\x2c\x02\x00\x16\x03\x56\x01\x05\x70\x75\x73\x68\x21\x16\x02\x01\x01\x2e\x01\x04\x62\x6f\x64\x79\x1a\x01\x04\x61\x72\x67\x73\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x59\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x03\x72\x65\x66\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x01\x2e\x01\x07\x43\x50\x79\x74\x68\x6f\x6e\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x1a\x2c\x5f\x00\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x1a\x2c\x61\x00\x16\x05\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x01\x07\x70\x79\x5f\x63\x61\x73\x74\x01\x02\x50\x79\x21\x03\x73\x79\x73\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x1a\x01\x07\x6d\x6f\x64\x75\x6c\x65\x73\x16\x05\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x56\x01\x07\x70\x79\x5f\x63\x61\x73\x74\x01\x02\x50\x79\x2c\x53\x00\x2c\x4f\x00\x17\xf6\x2c\x02\x00\x16\x02\x56\x01\x03\x65\x73\x63\x01\x03\x6f\x75\x74"
     sign_data = SHA.sha224(data)
     if sign_data != expected_sign
         error("The compiled data file", "E:\\Projects\\Git4\\jnumpy\\TyPython\\src\\CPython.Dev.DevOnly.compiled.jl", "is not verified. Please recompile it.")
