+++ TyPython/src/Utils.DevOnly.compiled.jl
@@ -1,0 +1,13 @@
+# This file is automatically generated by DevOnly.jl
+import Serialization
+import SHA
+let
+    expected_sign = b"\x08\xc8\x3e\x98\xb4\x2c\xe7\xe6\xb6\x43\x53\x07\x2d\xc3\xab\x84\xba\xde\xac\x31\x74\x3f\x61\xa6\x19\x24\x64\x75"
+    data = b"\x37\x4a\x4c\x0d\x04\x00\x00\x00\x16\x02\x01\x08\x74\x6f\x70\x6c\x65\x76\x65\x6c\x17\xe0\x01\x35\x45\x3a\x5c\x50\x72\x6f\x6a\x65\x63\x74\x73\x5c\x47\x69\x74\x34\x5c\x6a\x6e\x75\x6d\x70\x79\x5c\x54\x79\x50\x79\x74\x68\x6f\x6e\x5c\x73\x72\x63\x5c\x55\x74\x69\x6c\x73\x2e\x44\x65\x76\x4f\x6e\x6c\x79\x2e\x6a\x6c\x16\x02\x01\x05\x6d\x61\x63\x72\x6f\x16\x02\x56\x01\x0e\x73\x75\x70\x70\x72\x65\x73\x73\x5f\x65\x72\x72\x6f\x72\x01\x02\x65\x78\x16\x03\x01\x05\x62\x6c\x6f\x63\x6b\x17\xe0\x2c\x02\x00\x17\xe1\x2c\x02\x00\x16\x05\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x05\x62\x6c\x6f\x63\x6b\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x16\x02\x59\x01\x0a\x6f\x6c\x64\x5f\x73\x74\x64\x65\x72\x72\x01\x06\x73\x74\x64\x65\x72\x72\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x16\x02\x59\x16\x02\xa4\x01\x02\x72\x64\x01\x02\x77\x72\x16\x01\x56\x01\x0f\x72\x65\x64\x69\x72\x65\x63\x74\x5f\x73\x74\x64\x65\x72\x72\x16\x06\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x03\x74\x72\x79\x16\x04\x56\x24\x1f\x4e\x9b\x44\x01\x05\x5f\x65\x78\x70\x72\x1a\x01\x05\x62\x6c\x6f\x63\x6b\x01\x0a\x5f\x5f\x73\x6f\x75\x72\x63\x65\x5f\x5f\x16\x02\x56\x01\x03\x65\x73\x63\x01\x02\x65\x78\x4c\x4c\x16\x01\x01\x07\x63\x6f\x70\x79\x61\x73\x74\x1a\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x56\x2c\x0f\x00\x2c\x0a\x00\x16\x02\x56\x01\x05\x63\x6c\x6f\x73\x65\x01\x02\x77\x72"
+    sign_data = SHA.sha224(data)
+    if sign_data != expected_sign
+        error("The compiled data file", "E:\\Projects\\Git4\\jnumpy\\TyPython\\src\\Utils.DevOnly.compiled.jl", "is not verified. Please recompile it.")
+    end
+    buf = IOBuffer(data)
+    Base.eval(@__MODULE__(), Serialization.deserialize(buf))
+end
