+++ TyPython/src/Utils.jl
@@ -1,6 +1,7 @@
 module Utils
+import TyPython: DevOnly
 import IOCapture
-import MacroTools
+DevOnly.@devonly import MacroTools
 
 function capture_out(f)
     capture = IOCapture.capture() do
@@ -17,18 +18,6 @@ end
     block
 end
 
-macro suppress_error(ex)
-    MacroTools.@q begin
-        old_stderr = stderr
-        rd, wr = redirect_stderr()
-        try
-            $__source__
-            $(esc(ex))
-        finally
-            redirect_stderr(old_stderr)
-            close(wr)
-        end
-    end
-end
+DevOnly.@staticinclude("Utils.DevOnly.jl")
 
 end
